+++ packages/perseus/src/widgets/interactive-graphs/graphs/components/movable-point.test.tsx
@@ -1,8 +1,9 @@
-import {render, screen} from "@testing-library/react";
+import Tooltip from "@khanacademy/wonder-blocks-tooltip";
+import {render} from "@testing-library/react";
 import * as MafsLibrary from "mafs";
 import React from "react";
 
-import * as ReducerGraphState from "../../reducer/use-graph-state";
+import * as ReducerGraphConfig from "../../reducer/use-graph-config";
 
 import {StyledMovablePoint} from "./movable-point";
 
@@ -15,76 +16,132 @@ jest.mock("mafs", () => {
     };
 });
 
+jest.mock("@khanacademy/wonder-blocks-tooltip", () => {
+    const originalModule = jest.requireActual(
+        "@khanacademy/wonder-blocks-tooltip",
+    );
+    return {
+        __esModule: true,
+        ...originalModule,
+        default: jest.fn(),
+    };
+});
+
+const TooltipMock = ({children}) => {
+    return children;
+};
+
 describe("StyledMovablePoint", () => {
-    let useGraphStateMock: jest.SpyInstance;
+    let useGraphConfigMock: jest.SpyInstance;
     let useMovableMock: jest.SpyInstance;
     const Mafs = MafsLibrary.Mafs;
-    const baseGraphStateContext = {
-        state: {
-            snapStep: 1,
-            range: [
-                [0, 0],
-                [1, 1],
-            ],
-            markings: "foo",
-        } as any,
-        dispatch: () => {},
-        graphOptions: {},
+    const baseGraphConfigContext = {
+        snapStep: 1,
+        range: [
+            [0, 0],
+            [1, 1],
+        ],
+        markings: "graph",
+        showTooltips: false,
     };
 
     beforeEach(() => {
-        useGraphStateMock = jest.spyOn(ReducerGraphState, "default");
+        useGraphConfigMock = jest.spyOn(ReducerGraphConfig, "default");
         useMovableMock = jest
             .spyOn(MafsLibrary, "useMovable")
             .mockReturnValue({dragging: false});
     });
 
     describe("Tooltip", () => {
+        const graphConfigContextWithTooltips = {
+            ...baseGraphConfigContext,
+            showTooltips: true,
+        };
+
         /*  NOTE: When a WonderBlocks Tooltip is added to an element, it isn't rendered with the element.
                   Instead, it is rendered elsewhere in the document (usually the bottom of the <body>),
                         and only when the user interacts with the element.
                   This makes checking for the Tooltip element difficult to manage in a test.
-                  However, the tooltip does add an "aria-describedby" attribute to the targeted element.
-                  Therefore, these tests infer that a Tooltip has been added by looking at the "aria-describedby" attribute.
+                  Therefore, these tests mock the Tooltip component and check if/how it is called.
          */
+        beforeEach(() => {
+            // @ts-expect-error // TS2339: Property mockImplementation does not exist on type typeof Tooltip
+            Tooltip.mockImplementation(TooltipMock);
+        });
+
         it("References a tooltip when option is indicated", () => {
-            const graphStateContext = {
-                ...baseGraphStateContext,
-                graphOptions: {showTooltips: true},
-            };
-            useGraphStateMock.mockReturnValue(graphStateContext);
+            useGraphConfigMock.mockReturnValue(graphConfigContextWithTooltips);
             render(
                 <Mafs width={200} height={200}>
                     <StyledMovablePoint point={[0, 0]} onMove={() => {}} />,
                 </Mafs>,
             );
-            expect(screen.getByTestId("movable-point")).toHaveAttribute(
-                "aria-describedby",
-                "uid-tooltip-0-aria-content",
-            );
+            expect(Tooltip).toHaveBeenCalled();
         });
 
         it("Does NOT reference a tooltip when option is 'false'", () => {
-            const graphStateContext = {
-                ...baseGraphStateContext,
-                graphOptions: {showTooltips: false},
-            };
-            useGraphStateMock.mockReturnValue(graphStateContext);
+            useGraphConfigMock.mockReturnValue(baseGraphConfigContext);
             render(
                 <Mafs width={200} height={200}>
                     <StyledMovablePoint point={[0, 0]} onMove={() => {}} />,
                 </Mafs>,
             );
-            expect(screen.getByTestId("movable-point")).not.toHaveAttribute(
-                "aria-describedby",
-                "uid-tooltip-0-aria-content",
+            expect(Tooltip).not.toHaveBeenCalled();
+        });
+
+        it("Defaults to a 'blue' background if no color value is provided", () => {
+            useGraphConfigMock.mockReturnValue(graphConfigContextWithTooltips);
+            render(
+                <Mafs width={200} height={200}>
+                    <StyledMovablePoint point={[0, 0]} onMove={() => {}} />,
+                </Mafs>,
+            );
+            // @ts-expect-error // TS2339: Property mock does not exist on type typeof Tooltip
+            expect(Tooltip.mock.calls[0][0]).toEqual(
+                expect.objectContaining({backgroundColor: "blue"}),
+            );
+        });
+
+        it("Uses the color NAME of the hexadecimal color passed to the point for the background", () => {
+            useGraphConfigMock.mockReturnValue(graphConfigContextWithTooltips);
+            render(
+                <Mafs width={200} height={200}>
+                    <StyledMovablePoint
+                        point={[0, 0]}
+                        color="#9059ff"
+                        onMove={() => {}}
+                    />
+                    ,
+                </Mafs>,
+            );
+            // @ts-expect-error // TS2339: Property mock does not exist on type typeof Tooltip
+            expect(Tooltip.mock.calls[0][0]).toEqual(
+                expect.objectContaining({backgroundColor: "purple"}),
+            );
+        });
+
+        it("Defaults to a 'blue' background if the color value provided doesn't match WB colors", () => {
+            useGraphConfigMock.mockReturnValue(graphConfigContextWithTooltips);
+            render(
+                <Mafs width={200} height={200}>
+                    <StyledMovablePoint
+                        point={[0, 0]}
+                        color="#f00"
+                        onMove={() => {}}
+                    />
+                    ,
+                </Mafs>,
+            );
+            // @ts-expect-error // TS2339: Property mock does not exist on type typeof Tooltip
+            expect(Tooltip.mock.calls[0][0]).toEqual(
+                expect.objectContaining({backgroundColor: "blue"}),
             );
         });
     });
 
     describe("Hairlines", () => {
         it("Shows hairlines when dragging and 'markings' are NOT set to 'none'", () => {
-            useGraphStateMock.mockReturnValue(baseGraphStateContext);
+            useGraphConfigMock.mockReturnValue(baseGraphConfigContext);
             useMovableMock.mockReturnValue({dragging: true});
             const {container} = render(
                 <Mafs width={200} height={200}>
@@ -101,7 +158,7 @@ describe("StyledMovablePoint", () => {
         });
 
         it("Hairlines do NOT show when not dragging", () => {
-            useGraphStateMock.mockReturnValue(baseGraphStateContext);
+            useGraphConfigMock.mockReturnValue(baseGraphConfigContext);
             const {container} = render(
                 <Mafs width={200} height={200}>
                     <StyledMovablePoint point={[0, 0]} onMove={() => {}} />,
@@ -114,9 +171,9 @@ describe("StyledMovablePoint", () => {
         });
 
         it("Hairlines do NOT show when 'markings' are set to 'none'", () => {
-            const graphStateContext = {...baseGraphStateContext};
-            graphStateContext.state.markings = "none";
-            useGraphStateMock.mockReturnValue(graphStateContext);
+            const graphStateContext = {...baseGraphConfigContext};
+            graphStateContext.markings = "none";
+            useGraphConfigMock.mockReturnValue(graphStateContext);
             useMovableMock.mockReturnValue({dragging: true});
             const {container} = render(
                 <Mafs width={200} height={200}>