+++ packages/perseus/src/widgets/interactive-graphs/graphs/segment.tsx
@@ -13,16 +13,14 @@ type SegmentProps = MafsGraphProps<SegmentGraphState>;
 
 export const SegmentGraph = (props: SegmentProps) => {
     const {dispatch} = props;
-    const {coords: segments, snapStep, range} = props.graphState;
+    const {coords: segments} = props.graphState;
 
     return (
         <>
             {segments?.map((segment, i) => (
                 <SegmentView
                     key={i}
-                    collinearPair={segment}
-                    snaps={snapStep}
-                    range={range}
+                    points={segment}
                     onMoveLine={(delta: vec.Vector2) => {
                         dispatch(moveLine(i, delta));
                     }}
@@ -44,7 +42,7 @@ export const SegmentGraph = (props: SegmentProps) => {
 const SegmentView = (props: InteractiveLineProps) => {
     const {
         onMoveLine: onMoveSegment,
-        collinearPair: [start, end],
+        points: [start, end],
     } = props;
 
     return (