+++ packages/perseus/src/widgets/interactive-graphs/mafs-graph.tsx
@@ -21,7 +21,7 @@ import {
     getGradableGraph,
     initializeGraphState,
 } from "./reducer/interactive-graph-state";
-import {GraphStateContext} from "./reducer/use-graph-state";
+import {GraphConfigContext} from "./reducer/use-graph-config";
 
 import type {InteractiveGraphState, InteractiveGraphProps} from "./types";
 import type {Widget} from "../../renderer";
@@ -124,11 +124,12 @@ export const MafsGraph = (props: MafsGraphProps) => {
     }, [dispatch, xMinRange, xMaxRange, yMinRange, yMaxRange]);
 
     return (
-        <GraphStateContext.Provider
+        <GraphConfigContext.Provider
             value={{
-                state,
-                dispatch,
-                graphOptions,
+                range: state.range,
+                snapStep: state.snapStep,
+                markings: props.markings,
+                showTooltips: !!props.showTooltips,
             }}
         >
             <View
@@ -190,6 +191,6 @@ export const MafsGraph = (props: MafsGraphProps) => {
                     </Mafs>
                 </View>
             </View>
-        </GraphStateContext.Provider>
+        </GraphConfigContext.Provider>
     );
 };