+++ testing/renderer-with-debug-ui.tsx
@@ -15,18 +15,22 @@ import {registerAllWidgetsForTesting} from "../packages/perseus/src/util/registe
 import SideBySide from "./side-by-side";
 
 import type {PerseusRenderer} from "../packages/perseus/src/perseus-types";
-import type {APIOptions} from "../packages/perseus/src/types";
+import type {ComponentProps} from "react";
 
 type Props = {
     question: PerseusRenderer;
-    apiOptions?: APIOptions;
-    reviewMode?: boolean;
-};
+} & Partial<
+    Omit<
+        ComponentProps<typeof Renderer>,
+        "content" | "images" | "widgets" | "problemNum"
+    >
+>;
 
 export const RendererWithDebugUI = ({
     question,
     apiOptions,
     reviewMode = false,
+    ...rest
 }: Props): React.ReactElement => {
     registerAllWidgetsForTesting();
     const ref = React.useRef<Renderer | null | undefined>(null);
@@ -65,6 +69,7 @@ export const RendererWithDebugUI = ({
                             problemNum={0}
                             apiOptions={{...apiOptions, isMobile}}
                             reviewMode={reviewMode}
+                            {...rest}
                         />
                     </View>
                     <View style={{flexDirection: "row", alignItems: "center"}}>